Add test for verifying catching up from Inactive Window, and a fix for the situation the test recreates.  (#1417)
* Add test for verifying catching up from Inactive Window.

The goal of this test is to verify full catch up of a Replica
only from the Inactive Window by isolating it from all others
except the Primary, so that State Transfer will not be able to
start.

* Fix for helping peers from Inactive Window.

The fix contains:
1. Transfer the SeqNumInfo completely. This way if slow path
   has started we can help from the Inactive Window.
2. Add resend of PrePrepares in Status Exchange mechanism in
   the Primary for the Inactive Window. Replicas will not ask
   for SeqNums in the timer onInfoRequestTimer unless they have
   the PrePrepare for the given SeqNo.

* Add wait for Stable Checkpoint collection in all Replicas.

Wait for replicas to collect Stable Checkpoint before
returning from this function. This way they will be ready
to help with State Transfer to peers that will be started
after this function returns.

* Run skvbc_chaotic_startup suite under root.

Because we have networking partitioning in this suite
which uses iptables, we need to run it as super user.

* Disable test for builds with BUILD_COMM_TCP_TLS=true.

* Readability improvement. Use of fill_and_wait_for_checkpoint
to avoid code duplication.