Improve noop mechanism on wedge command (#1431)
The former implementation of the noop mechanism was as follows:
Once a primary replica has detected that we need to wedge on seq num X it sent empty prepreapres up to X.
This created several problems:

It violates the concurrency level - creating many slow path commits
No new blocks were written for these sequence numbers, this can damage the sync mechanism in case of a crash
In this PR we introduce a new mechanism:

Once the primary done executing a request, and it detects that we need to wedge on seqNum X, it uses the internal client to send a special noop reconfiguration request. This noop request is executed by the reconfiguration engine which means that it also creates a new block.
In addition, now, we send a new prepreapre (containing the noop) only after executing one - which means we don't violate the concurrency level principle.
In the future, we plan to add a verification step to validate this noop request. However, even now, it does not violate anything, if a malicious client wants to launch a DOS attack, it can also send regular heavy computed requests (rather than noops)