Enable ignoring the primary in the matched part of bft replies (#1423)
The read-only replica does not know who the primary is. However, when matching a reply we also look at the primary from the replica.
To enable reading from read-only replicas, we added another configuration option to the matcher that instructs him to ignore the primary. This way, only if the user gives a configuration that contains ro replicas, we ignore from the primary in the replies

Notice that in this case, the current primary of the client won't be affected