Close socket gracefully when MetricServer is stopped
Socket close invalidates the fd while connection is still active on the other thread. This can cause race condition because same fd is in use by system calls in another thread.  Call shutdown first to close the active connections associated with the fd. This causes recvfrom to return and socket is closed gracefully in the same thread.

Reduced timeout to 1 second, to cause receive thread to timeout sooner in case recvfrom does not return after socket shutdown.