Fix inactive window catchup up with gap (#1472)
* Added test for catching up from Inactive Window with gap.

In this test we set things up so that one Replica will be behind
the peers by 3 Checkpoints on the Last Stable SeqNo, but only 2
Checkpoints behind in the Lase Executed. This recreates a problem
with Replica Status exchange mechanism.

* Fix for searching in Inactive Window on Status exchange.

Replicas helping from Inactive Window will access only the related SeqNo-s.