Implement framework for n-f reconfiguration actions + specific implementation for n-f wedge (#1461)
In this PR we present a framework for supporting n-f reconfiguration actions.
Until now, the reconfiguration engine has worked under the assumption that we have a n/n quorum. However, in many cases, this assumption does not hold.
To support n-f reconfiguration actions we add a dedicated reconfiguration SM that implements a callback for state transfer.
Once state transfer is completed, the call back is being called and we can change the reconfiguration state accordingly.
Notice, that the above is supported by the fact that we write every reconfiguration command to the blockchain.

In addition, in this PR we present an n-f wedge implementation.
To implement n-f wedge, we did the following:

A replica reports on a wedge on a stable checkpoint already (instead of n/n checkpoint)
In addition, once a replica is starting, it unwedge if necessary (this is a temporary solution until we implement a real unwedge request)
Thus, a replica that is in ST has two options: (a) it done with ST before the replicas have restarted. In this case, once ST is done, the data is in the reserved pages and the replica simply behaves like all other replicas. (b) The replica completed the ST after the other replicas have unwedged, in this case, we need to do nothing and simply continue with the quorum.
To validate the n-f wedge, we added an Apollo test for this case.
Note that it is still possible to wait for n/n wedge.