Clang-tidy: Ignore protobuf/gRPC generated header files
We went out of our way to ignore source files via .clang-tidy-ignore in
previous commits but this isn't good enough for header files. The latter are
treated separately and if any of your non-ignored source files includes a
header file that you don't want to check then we need to adapt .clang-tidy.
This change ignores any reports for protobuf/gRPC generated header files.