Clang-tidy: Remove clang-tidy workaround in CMake
With the prior clang-tidy changes in place, we don't need the workaround in
CMake anymore. It is a workaround because we overwrite clang-tidy's
configuration to only do a single/simple test. However, we don't want any
checks to be reported for some files. Now, we have .clang-tidy-ignore which
ignores source files for real.