State transfer in separate thread (#1488)
* Enabling State Transfer in Separate thread by default in tester replica

We have the option to run state transfer mechanism in a separate thread from BFT message processing. Running ST in a separate thread has become an established way for the system in real deployments. So enabling option to run ST in separate thread.

* Enabling State Transfer in Separate thread by default in tester replica

We have the option to run state transfer mechanism in a separate thread from BFT message processing. Running ST in a separate thread has become an established way for the system in real deployments. So enabling option to run ST in separate thread.

* Enabling State Transfer in Separate thread by default in tester replica

We have the option to run state transfer mechanism in a separate thread from BFT message processing. Running ST in a separate thread has become an established way for the system in real deployments. So enabling option to run ST in separate thread.

* Resolving as per comments

Co-authored-by: utkarsh katiyar <ukatiyar@vmware.com>