Improve and refactor the reconfiguration handler (#1479)
In this PR we present a couple of refactoring to the reconfiguration mechanism:

We united the IPruningHandler and the IReconifugrationHandler. Now we have only IReconfigurationHandler
The above means that pruning handlers are now considered reconfiguration handlers
We simplified the dispatcher code to use visitor instead of std::holds_alternative
Each reconfiguration Handler handles only the requests it can verify. This allows us to have more flexibility with the responsibilities chain of reconfiguration
We added a basic bftReconfigurationHandler that is responsible for verifying the operator requests
We separated the internal reconfiguration handler (that is now responsible for writing the wedge noops to the blockchain) from the regular reconfiguration handlers, giving it its own validation mechanism