Add db_categories.h to concord-bft
This commit copies the categories defined in categorized_storage.hpp
and the `kConcordInternalCategoryId` defined in db_interfaces.h to
kvbc/include/categorization/db_categories.h in concord-bft .
The motivation behind this change is that the value of
`const auto kExecutionEventsCategory` is used in kvbc_app_filter.cpp to
check if the blocks have events or not. We consolidate all the
categories together to accommodate this dependency into
db_categories.h.
Please note that this MR doesn't update the usage of
`kConcordInternalCategoryId` in the repository, since it would be
better to do that for both closed-source and open-source repo
at the same time to avoid conflicts.