ConcordClient: Interface
With this change we introduce the concord_client.hpp API. The main focus is on
usabilitty. The interface has three function calls - two for sending a write
and read request and a third to subscribe to events from the replica nodes.
All three APIs take a templated callback as a parameter which gives the user
the most flexibility. However, callbacks mean that we rely on the user to do
expensive computation in another context. This is a conscious trade-off.
Note: The interface isn't backed by an implementation yet and is supposed to
serve as a guideline for the clientpool and thin replica client interface.
With that, we expect minor changes to the interface going forward.