Persist all Checkpoint msgs that serve as proof for Replica's lastStableSeqNum (#1510)
* Persist all Checkpoint msgs that serve as proof for Replica's lastStableSeqNum

After adding signatures to checkpoint messages in order for Replicas to be
able to forward them we need to also persist those messages. This way after
restart we still will be able to prove to the Peers that our lastStableSeqNum
really is stable.

* Added unit test for seralizing/deserializing CheckpointMsg-s.

In this test we create Checkpoint messages to serialize them
and then deserialize back to check that they were recovered
in the same way they were when we initially created them.