Fix a few bugs in the diagnostics_server
It's possible that if the readline function, including read itself takes
longer than a second, that the remaining time can become negative and
cause select to fail in the next iteration of the loop. We prevent that
with a check at the bottom of the readline function.

Pass `errno` to `errnoString` instead of the return value from
failing libc calls.

Use the correct type for the return value of `read` and also for `count`.