Cherrypicking race-condition fixes from master branch to v0.11.x (#1567)
* This PR fixes a race-condition observed in reading and writing an unordered_map.

In favour of performance, have avoided using locks. Instead, I have made value part of unordered_map as atomic so that reader thread doesn't read a stale value while writer has updated the value for that key.

* This PR fixes a race condition observed in accessing variable beginningOfActiveWindow from two threads.

I have made this variable conditional atomic as it is not required to be atomic in all cases, only for the case reported using thread sanitizer. In our case only mainLog template instantation is the culprit.

Co-authored-by: utkarsh katiyar <ukatiyar@vmware.com>