ClientService: Create ClientService wrapper
This change introduces the ClientService class which is used by our main()
function to start the gRPC server. It encapsulates both gRPC services, the gRPC
server, and in the future the ConcordClient and most likely a ClientService
configuration. The latter two will need to be initialized outside the
ClientService. Thereby, we connect all parts through composition.