Logging: Change default format
This change aligns with the closed-source logging format. The main motivation
is to print the timestamp first so that messages can be sorted by the order in
which they appear. Having the replcia ID in the front makes sense when all
replicas print to the same destination. However, that is not the case and when
used in a docker-compose envrionment then it gets prefixed with the container
id anyways.