Persist the last KVBC block ID in metadata
Perist the last KVBC block ID in metadata as part of the effort to
change replica state sync so that we no longer persist the last executed
BFT sequence number in KVBC blocks.

Introduce an user data field in metadata as a
`ConstMetadataParameterIds` type. Reduce `reservedSimpleParamsNum` to
ensure adding user data is backwards compatible.

Persist the last block ID in metadata in two cases:
 * once request processing has completed, in
   `IRequestsHandler::onFinishExecutingReadWriteRequests()`.
 * once state transfer completes, potentially adding new blocks

A subsequent commit will change replica state sync so that it uses the
block ID instead of the BFT sequence number.

Improve const-correctness in MetadataStorage as part of this commit.