Define CMF messages for SKVBC
This commit defines CMF messages intended for use with SKVBC (i.e.
simpleKVBC, implemented in tests/simpleKVBC) for communication between
the SKVBC clients and replicas.

The intention is to eventually move SKVBC clients and replicas
(including those implemented in tests/simpleKVBC/TesterClient and
tests/simpleKVBC/TesterReplica and also Apollo's usage of SKVBC
implemented in tests/apollo/util/skvbc.py) to use these messages
defined in CMF rather than their current ad-hoc message format rooted
in manually packing C++ struct and object data. We hope the switch to
CMF in this area will improve the code cleanliness, readability, and
long-term maintainability.

If this commit is accepted, it will be followed up at a later time with
changes to actually switch the SKVBC clients and replicas (as outlined
above) to use the messages newly defined in this commit.