ClientService: Document event stream limitation
At the moment, the thin replica client only supports a single subscription at a
time. Given that concord client will be backed by the TRC this limitation will
extend to the gRPC interface. Ofc, this is not intuitive for a gRPC service but
we hope to correct this in the future.