Introduce a reserved pages client for pruning
Provide persistence of pruning agreement and progress information in
reserved pages. Pseudo code for how it can be used is provided as an
inline comment.

Introduce the `ReservedPagesMock` class and reuse it in ccron and
pruning reserved pages client unit tests. In order to make it work, add
the static `getNumResPages()` method and make the `my_offset()` and
`calc_my_offset()` methods static instead of const in `ResPagesClient`.
That allows for a simpler init procedure in unit tests as we can now
call them without an instance.

Move the `INITIAL_GENESIS_BLOCK_ID` constant to kv_types.hpp and reuse
it across the codebase.

Add an unit test to verify the pruning reserved pages client's
functionality.