PreProcessedResultMsg implementation
PreProcessedResultMsg is used by the PreProcessor. It is inserted into
the primary's message queue once a client request has been successfully
preprocessed by the replicas.

The message inherits from ClientRequestMsg mainly because it needs to be
bundled with ClientRequestMsg in PrePrepare. Code reuse is a side benefit,
but not the main reason for the inheritance.