BCStateTran: add historgams,metrics,DurationTracker,refactor code
This commit includes the following changes:

1) Log state transitions and ,odify statistics during calls to
getFetchingState(). Every time a state changes, a call to a new function
onFetchingStateChange() is made. This function modifies duration
trackers, Throughput objects and other statistics members.
The last state is saved into lastFetchingState_.

2) Add metrics counters for all state trandfer messages.
Add histograms to particular ones.

3) For each cycle, log first and the last block required and all
source ids.

4) Change Throughput report window getMissingBlocksSummaryWindowSize to
4. That means, a report is made now every 600 blocks collected
(was 150).

5) Convert all previous duration measurements, and add new ones.
In total, we now have 6 duration trackers, some might be removed later.
All duration trackers will be included in the cycle summary
(next commit).

6) reportCollectingStatus - add an option to trigger log report which is
not automatic, and not on the end of the throughput window. This is done by
adding a new parameter toLog. This allows us to get the last report when
the cycle ends.

7) historams - divide into 3 components: state_transfer,
state_transfer_src, state_transfer_dest. Add new histograms,recorders
and async redocrders to get statistics of various operations durations,
queue and message sizes.

8) Clarify some logs, add extra info to them.

9) Remove unneeded code and comments.

10) Fix typo throughtput -> throughput in some variable names.

11) histograms (dest) - snapshot every cycle beginning,
to reset histogram.