BCStateTran: add histogram snapshots to log for source replica
Until now, we used to log snapshots only for destination, at the end of
each cycle. As source replica receives a series FetchblocksMsg message.
We would like to log statistics for this period of time. Here we track
the source selection, raise a flag sourceFlag_ 1st time a source
receives FetchblocksMsg. When the source stops receiving FetchblocksMsg
messages for some time, using a counter, we snapshot the histograms
and stringify them to log.