ConcordClient: std::function instead of templates
At the moment, we don't see the callback invocation itself as a performance
bottleneck. For both APIs, send() and subscribe(), concord client is going to
wait on multiple replicas to respond and compare the results. Invoking the
callback via templates or std::function afterwards seems negligible in
comparison. However, we also have no data (we should do this once concord
client is fully implemented). We (@robem and @f-squirrel) agreed that
std::function makes the interface straightforward and easier to maintain.
Templates require all functions to be defined in header files and callback
types to be "defined" in comments. If performance is not an issue then we
prefer a clean interface via std::function.