ConcordClient: Offer metrics API
In order to collect metrics in concord client and its subcomponents, we are
using our own basic metrics framework. We didn't want to depend on another 3rd
party dependency but the user can still choose to use Prometheus,
OpenTelemetry, ... on top of it.
The actual usage of the Aggregator will come with the integration of the client
pool and the thin replica client.