BCStateTran: snapshot histograms and log in source and destination
- We need to snapshot histograms at the beginning of a measured
interval, and at the end (for logging purposes)
- We now have 3 components. For source - log "state_transfer_src" and
"state_transfer", for destination log "state_transfer_dest" and
"state_transfer".
- To track source time interval - we add an algorithm to track source
selection and to snapshot histograms when we assume that source is no
longer source. sourceSnapshotCounter_ is incremented every time the
refresh timer is triggered. This is strong evidence that the source is
not sourced again. If the time passed using this counter is roughly
twice than fetchRetransmissionTimeoutMs, we declare that source is
no longer source and snapshot histograms to log.
- The destination histograms are snapshot at the beginning of the cycle
(reset) and at the end of GettingMissingblocks - This is a more accurate
time to snapshot.