Introducing Client reconfiguration engine (#1626)
In this PR we present the Client reconfiguration engine (CRE)
The cre is composed of three pieces:

IStateClient - a mechanism to get the next state update
IStateHandler - set of handlers that each one is responsible for executing a single update. In addition, if needed, they also know how to update the client state on the chain (using a bft write commands)
ClientReconfigutaionEngine - a component that loops while and waits for reconfiguration updates.
In this PR we present the very basic interfaces and implementation on the CRE side.

Interfaces
Poll based state client
unit tests for the API as well as the poll-based state client implementation.