State Transfer: disable sourceReplicaReplacementTimeoutMs
Now that we can disable source peridic replacment (see previous commit),
we are going to use it.
This is done by setting sourceReplicaReplacementTimeoutMs=0.
We do it due to the fact that source-selector chooses to replace a
source in the middle of a batch.
This reduces throughput by 10-20%. For now, we choose to disable it.
This feature will be re-designed and re-enabled in the near future.