IAppState: add startGetBlockAsync::getBlockAsync
Here we add a virtual interface for IAppState.
This allows getting a blocks asynchronously. In practice, we will use
this new call to invoke multiple threads to get blocks concurrently.

In order to implement this feature in class Replica (which inherits
IAppState), concord::util::ThreadPool is used. The default size of this
pool equal to the number of CPUs when set to 0.