ClientService: Parse configuration file
Our closed source clients parse a YAML configuration file at runtime. With this
change, we will parse the same file in the clientservice and use the values to
fill ConcordClient's configuration structures. This is by no means complete but
a start so that we can continue to fill in the blanks.