Client reconfiguration engine - complete implementation on replica side  (#1660)
in this PR we complete the CRE implementation on the replica side.

We maintain a state on the blockchain per CRE id (which is eventually a client node id)
We write and read the relevant data per CRE requests
To enable the above we made some more changes:

The sender ID was added to the reconfiguration request and to the validation method (in order to allow validation per client)
We allow self-verification in SigManager
To improve the CRE functionality each handler may return a WriteState which will be written to the blockchain. This WriteState also contains a callback that is invoked after the write command has succeeded. This allows us to perform actions write after updating the new state for the client (for example, after writing a new public key to the replica, rotate the private key).

Finally, we added a basic Apollo test that demonstrates the capabilities of CRE. By running the test, one can see in the logs a simulation of client key-exchange using CRE

To complete the implementation we need to decide how to verify the client reconfiguration requests. This can be either by using the transactions signing key or by having a dedicated key for CRE