Byzantine Primary for PRA (#1672)
* Byzantine Primary for PRA

PreExecution Result Authentication require a
Byzantine Primary Replica.
This feature should work even if the Byzantine
Primary Replica is present in the network.All
the apollo test suite goes into view change and
then re-elect a new primary due to this feature.
If this feature is off, the same Byzantine Replica
is creating a deferred consensus. Due to this
phenomenon the test will not stop as clients will
wait for the result for a long time.

A new strategy is created for this Byzantine
Replica.

* Code review comments updation

Updating the comments given in the code review.