TRS: Update metrics aggregator periodically in SubscribeToUpdates while loop
This commit adds update_metrics_aggregator_thresh as a TRS config parameter.
The value of the parameter defines the frequency with which metrics
aggregator is updated. We do not call updataAggregator everytime in the
`SubscribeToUpdates` while loop because it is an expensive operation, that takes
a lock and then copies metrics accross threads.