Clientservice: Fix subscriber thread handling
So far we used the "stop notifaction" variable to check whether the
subscription thread is still running. Well, this is error prone given that the
notification change itself doesn't reflect the actual thread termination. This
PR wraps the thread in a pointer which allows us to check easily whether the
subscription worker is still active.