Introducing epoch manager (#1679)
In this PR we introduce the epoch manager and demonstrate its power via add/remove node command with lagged replicas.
Epoch is a life cycle of the system. by changing epoch we indicate that we actually started a new system on top of an old state.
All replicas have to decide on starting a new epoch exactly at the same state to preserve consistency. Thus, starting a new epoch after wedge makes great sense.
Practically, when a replica decides to start a new epoch, it writes a new "genesis" block that contains this new epoch under a pre-known key.