Bftclient: Fix build dependency
All CMake targets that use bftclient_new most likely link directly or
indirectly to secretsmanager. The yaml_parsing test in this PR doesn't do this
and hence exposes the missing dependency.
The same seems to be true for threshsign but there is more to do and hence this
change fixes the problem at hand only.