ConcordClient: Legacy events in proto
In order to support the events API that existed before concord client, we
introduce Events in concord_client.proto as a possible value to be streamed to
the caller. Essentially, events and event groups are different formats on the
blockchain and therefore, the caller might retrieve both types. However, we're
switching to the latter and hence, once the stream turned to EventGroups it
won't go back to Events.