Adding reconfiguration support for RO Replica
In this PR, we add changes to addRemove reconfiguration command with epoch number in reserved pages.
When starting a new epoch, check if there is reconfiguration command in the chain, if yes, add it to the reserved pages, so that CRE in RO replica can get this command and act on it.