ClientPool: Clang-tidy fixes
Clang-tidy complains about unused namespaces and variables that should be moved
or used as references. Those are easy adjustments. The more complicated one is
in `SendRequest()`. Clang-tidy complains that `request` is exposed to a
use-after-move issue. Practically, this cannot happen due to the boolean which
marks whether the request was processed (moved). However, clang-tidy cannot
reason about this and TBH, it is not obvious right away. Instead of suppressing
this warning, we can refactor the function to return early whenever a request
got processed.