IAppState:: add argument lastBlock flag in putBlock()
This flag is defaulted to true. So every time the caller put a block,
definition (see Replica.cpp) will try to link between the temporary
chain and the blockchain. For this, there is additional processing
and IO operations. The caller knows that the actual linkage would happen
only on the last call to putblock(), so it can choose
(for an improved performance) to call all calls till the last one  with
this new flag set to false, while in the last one calling with the flag
set to true.