IAppState: add new interface overloaded function getPrevDigestFromBlock
This is an overload of getPrevDigestFromBlock.
We will soon need it in one of the next commits, to extract the digest
straight from the raw block arriving from the source replica during
state transfer. We are planning to put blocks concurrently, or in the
more fat future collects them in a different order. when putting
concurrently, the parent block might not have been committed yet, so the
current call to the existing getPrevDigestFromBlock will fail - the
implementation relays on the fact that parent block has already been
committed to storage, which is not true when committing multiple blocks
concurrently.