IAppState: add maximal block size to getBlock and getBlockAsync
This allows detecting issues on spot, instead of having a potential
buffer overflow. There is a future task to re-design these specific
API functions, but for now this is the minimum we can do.

Also, do not catch any exceptions. These can be catched by caller when
calling future.get().