TRS: Pass block_id associated with event_group_id as syncAndSend arg
This commit identifies the block id associated with the requested
event_group_id if the subscription request is an event_group request.
This block id is used as the start_block_id in the call to
syncAndSend method. This change is required because we cannot query
the blockchain directly using an event_group_id.