Update kvbAppFilter to get event groups
This commit adds a method `getEventGroup` that retrieves
the event_group associated with the event_group_id in kv_blockchain.
EventGroups are serialized and stored in ImmutableInput as
a map of {event_group_id, serialized_event_group}. The event_groups
are retrieved based on their event_group_id, deserialized and
returned.
Methods corresponding to ReadBlockRange, ReadBlockHash and
ReadBlockRangeHash methods are added for event_groups, similarly
filterEventsInEventGroups is added that filters events in
the event_groups based on the event tags.