Clientservice: Connect to TRS over insecure channel
This change makes the TRC actually connect with the replicas' TRS service.
In addition, the subscription struct has been changed given that we do "all or
nothing" for TLS anyways. However, this needs to be revist when we support TLS
for TRC-TRS connections. Note, the concord client interface requires that the
certificates have been read into a buffer but the current TRC interface would
like to recieve a file path. The former is the desired outcome but the latter
could be used for easy integration. A later change will take care of this.