TRC: Support Event Groups
This big change adds event groups, which are received from the TRS, to the
Update structure which is send to TRC callers. Besides adjusting mostly
everything with if-else statements, we change the way hashes are stored for our
internal f+1 comparison. Instead of using `std::pair` as key to the map, this
change adds a struct and hence a custom comparison function for `std::map`.