Remove acquire methods as they are no longer needed by SequenceWithActiveWindow.
Previously we had 2 circular buffers, one for the Active Window and one
for the Inactive Window. This required correct transferring of the content
of the Elements stored in those circular buffers. After the refactor
We store both the Active and the Inactive parts in one circular buffer and
those functions are no longer needed.