Clientservice: QoL YAML reading
Exceptions from yaml-cpp are quite ambiguous. This change catches any YAML
exception and prints the index it tried to read from the given YAML node. In
addition, if default values are used then we log this explicitly to avoid any
confusion at runtime. I chose to not log the value given that not every value
is "printable" - there are ways around it but this change doesn't bother with
it. The fact that we inform the user about using default values seems valuable
enough at this point.