utils: introduce new utility SimpleMemoryPool, add unittest
This is a template class that implements a simple memory pool for a
single type of element.
A common use case for this class is the pre-allocation of N elements at
the start. This is done in order to:
1) Make sure the only possible allocation failures in the module can
happen on boot (startup).
2) Avoid allocation/deallocation of memory. Specifically, for small
memory chunks, in the long term this can cause
performance issues and memory defragmentation.
Currently, the class does not implements pool of elements of pointer or
array type.

Here we also add a gtest unit-test for the SimpleMemoryPool.