Support callbacks on consuming external messages
Add a functionality to **IncomingMsgsStorage** that allows users to pass
in a callback that will be called when the given message is consumed by
the replica. Such a functionality can be useful to be able to determine
if a particular message is still pending or not. An example is Concord
Cron's TicksGenerator - it will be able to know whether a tick is
pending in the external message queue.

Make pushExternal* methods return a boolean that is true on success and
false otherwise (e.g. queue capacity has been reached).

To implement, in addition to the unique_ptr to the message, store a
std::function callback, making the queue contain a std:pair of these
two. The resulting std::pair will remain movable only. Expectation is
that storing and moving an empty std::function will not degrade
performance noticeably.

Make sure we don't register the performance component if already
registered with the set of shared recorders. That allows us to have
multiple instances of **IncomingMsgsStorageImp**.

Add an unit test that mostly covers new functionality. More tests for
existing functionality will be added in separate commits.